debugged the hell out of the buffering RR performance today...

+++ apps/generic_apps/pltt_app/PLTT_passive_app.cpp
@@ -31,7 +31,7 @@ typedef Os::Debug Debug;
 typedef Os::Rand Rand;
 typedef Os::Timer Timer;
 typedef Os::Clock Clock;
-typedef uint8 TimesNumber;
+typedef uint16_t TimesNumber;
 typedef uint8 SecondsNumber;
 typedef uint32 AgentID;
 typedef wiselib::NeighborDiscovery_Type<Os, Radio, Clock, Timer, Rand, Debug> NeighborDiscovery;
